pyproject.toml
[project]
name = "uv-fasthtml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "python-dotenv>=1.1.1",
    "python-fasthtml>=0.12.21",
    "sqlmodel>=0.0.37",
]

[tool.pytest.ini_options]
pythonpath = [
    ".", "app"
]

[tool.ruff]
exclude = [
    "**/__pycache__",
    "migrations/"
]

[tool.ruff.lint]
ignore = [
    "E701",
    "F403", # Star imports
    "F405"  # Star imports
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-playwright>=0.7.1",
    "ruff>=0.13.2",
]
